calendar: always emit "day-selected" once
"day-selected" was emitted twice when clicking on a day of the previous
or next month. This was because calendar_set_month{next,prev}() always
selected the first (or last) day of the newly focussed month. Callers
that wanted a different day to be selected had to do so after calling
one of those functions.

Fix this by passing the desired day into those functions directly.

https://bugzilla.gnome.org/show_bug.cgi?id=755941