Allow setting menu item accelerator text directly
Add a custom GMenuItem attribute called "x-canonical-accel" that
allows the accelerator text to be set directly without needing a
key code and modifier. We need this because sometimes all we have
is the translated accelerator text, and parsing this is difficult
considering the user's language may not be English.
Author: William Hua <william.hua@canonical.com>
Bug-Ubuntu: https://bugs.launchpad.net/bugs/1208019